[Refactor] Move `Contents` to `re_viewer_context`, rename `SpaceViewName` to `ContentsName`, and move it to `re_viewer_context` as well
