+++ crates/re_space_view/src/lib.rs
@@ -15,7 +15,7 @@ mod visualizable;
 pub use data_query::{DataQuery, EntityOverrideContext, PropertyResolver};
 pub use heuristics::suggest_space_view_for_each_entity;
 pub use screenshot::ScreenshotMode;
-pub use space_view::{SpaceViewBlueprint, SpaceViewName};
+pub use space_view::SpaceViewBlueprint;
 pub use space_view_contents::SpaceViewContents;
 pub use sub_archetypes::{
     entity_path_for_space_view_sub_archetype, query_space_view_sub_archetype,
