+++ crates/re_viewport/src/context_menu/actions/remove.rs
@@ -1,9 +1,8 @@
 use re_entity_db::InstancePath;
 use re_log_types::EntityPathRule;
-use re_viewer_context::{ContainerId, Item, SpaceViewId};
+use re_viewer_context::{ContainerId, Contents, Item, SpaceViewId};
 
 use crate::context_menu::{ContextMenuAction, ContextMenuContext};
-use crate::Contents;
 
 /// Remove a container, space view, or data result.
 pub(crate) struct RemoveAction;
