+++ crates/re_viewport/src/lib.rs
@@ -24,7 +24,7 @@ mod viewport_blueprint_ui;
 /// Unstable. Used for the ongoing blueprint experimentation.
 pub mod blueprint;
 
-pub use container::{ContainerBlueprint, Contents};
+pub use container::ContainerBlueprint;
 pub use context_menu::{context_menu_ui_for_item, SelectionUpdateBehavior};
 pub use viewport::{Viewport, ViewportState};
 pub use viewport_blueprint::ViewportBlueprint;
